🚚 [projects] Rename module `{project => }config`
