+++ src/cutty/projects/generator.py
@@ -7,9 +7,9 @@ from dataclasses import dataclass
 
 from cutty.filesystems.domain.path import Path
 from cutty.filesystems.domain.purepath import PurePath
+from cutty.projects.config import createprojectconfigfile
+from cutty.projects.config import ProjectConfig
 from cutty.projects.project import Project
-from cutty.projects.projectconfig import createprojectconfigfile
-from cutty.projects.projectconfig import ProjectConfig
 from cutty.projects.template import Template
 from cutty.templates.adapters.cookiecutter.config import findcookiecutterhooks
 from cutty.templates.adapters.cookiecutter.config import findcookiecutterpaths
